Update to compile on OS X Mountain Lion - Created new project file (can't automatically upgrade from Project Builder to Xcode anymore) - Resolved minor code problems pointed to by Xcode (implicit conversion, deprecated APIs, etc) - Added check for nil value if Chicago font isn't present (it's no long included in OS X by default) - Updated to use ARC

+++ MissileView.h
@@ -110,10 +110,10 @@ typedef struct {
 	MissileStruct missile[maxBombs];
 	
 	// Current player's score
-	int score;
+	long score;
 	
 	// Current high score
-	int highScore;
+	long highScore;
 	
 	// How many loops have passed without a launch
 	int loopsWithoutLaunch;
